move static data into data folder
