+++ components/data/OfficersArray.js
@@ -55,7 +55,7 @@ export const OfficersArray = [
 		position: "Industry",
 		image: "samarth.png",
 		email: "ssrin027@ucr.edu",
-		linkedin: "www.linkedin.com/in/samarth-srinivasa",
+		linkedin: "https://www.linkedin.com/in/samarth-srinivasa",
 		github: "https://github.com/SamarthSrinivasa",
 		bio: "I am a third year data science major at UCR! Contact me if you have any questions about talking to industry leaders and companies in the field of computer science, or any other questions you might have about ACM! Other than School and ACM, I enjoy playing music, chess, and cooking.",
 	},
@@ -75,7 +75,7 @@ export const OfficersArray = [
 		position: "Marketing",
 		image: "ashley.jpg",
 		email: "akim291@ucr.edu",
-		linkedin: "www.linkedin.com/in/ashley-kim-838893196",
+		linkedin: "https://www.linkedin.com/in/ashley-kim-838893196",
 		github: "",
 		bio: `Hi everyone! I'm Ashley and I'm a 3rd year CS major. My hobbies outside of class include digital illustration, making fashion pinterest boards, and baking cookies for my friends.`,
 	},
@@ -85,7 +85,7 @@ export const OfficersArray = [
 		position: "Marketing",
 		image: "ashwin.jpg",
 		email: "ajaga008@ucr.edu",
-		linkedin: "www.linkedin.com/in/ashwin-jagadish",
+		linkedin: "https://www.linkedin.com/in/ashwin-jagadish",
 		github: "",
 		bio: `Hi, my name is Ashwin, and I'm a 3rd year Computer Science with Business Applications major. I'm looking forward to being ACM's Marketing Chair this year and creating more engagement within the club.`,
 	},
